furniture_delivery/requirements.txt
allure-python-commons==2.9.45
asgiref==3.5.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.3
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.1
django-allauth==0.50.0
django-coverage-plugin==2.0.3
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-test-plus==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
model-bakery==1.5.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-sugar==0.9.4
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
